updated sets to use service and segregated models
